+++ WS-ui/src/main/java/org/superb/apps/ws/db/functionalities/CRUDInterface.java
@@ -17,7 +17,7 @@ public interface CRUDInterface<T> {
     //<editor-fold defaultstate="collapsed" desc="data to read">
     public List<T> getAll();
 
-    public T getByID(int ID);
+    public T getByID(long ID);
 
     public List<T> getByName(String partialName);
     //</editor-fold>
